+++ resources/src/js/store/modules/base/getters.js
@@ -1,0 +1,27 @@
+export default {
+    courses: state => state.courses,
+    roster: state => state.roster,
+    flashroster: state => state.flashroster,
+    list: state => state.list,
+    flash: state => state.flash,
+    menushow: state => state.menushow,
+    courseid: state => state.courseid,
+    courseTitle: state => state.courseTitle,
+
+    faculty_email: state => state.faculty_email,
+    faculty_name: state => state.faculty_name,
+    faculty_profile: state => state.faculty_profile,
+    faculty_first_name: state => state.faculty_first_name,
+    faculty_last_name: state => state.faculty_last_name,
+    faculty_full_name: state => state.faculty_full_name,
+    faculty_profile_image: state => state.faculty_profile_image,
+
+    sp_student_id : state => state.sp_student_id,
+    sp_emailURI : state => state.sp_emailURI,
+    sp_display_name: state => state.sp_display_name,
+    sp_major: state => state.sp_major,
+    sp_bio: state => state.sp_bio,
+    sp_images: state => state.sp_images,
+    sp_image_priority: state => state.sp_image_priority,
+    sp_notes: state => state.sp_notes,
+}
