+++ resources/src/js/store/modules/game/actions.js
@@ -1,0 +1,5 @@
+export default {
+    dummyAction(context) {
+        context.commit('DUMMY_MUTATION');
+    }
+}
