+++ resources/src/js/store/modules/game/getters.js
@@ -1,0 +1,3 @@
+export default {
+    test: state => state.test
+}
