+++ resources/src/js/store/modules/game/mutations.js
@@ -1,0 +1,5 @@
+export default {
+    DUMMY_MUTATION(state) {
+        state.test = true;
+    }
+}
