+++ resources/src/js/store/modules/game/state.js
@@ -1,0 +1,3 @@
+export default {
+    test: true
+}
